rewrite cogs.tools_cog.py file into separate cogs
